pyproject.toml
[project]
name = "test-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0",
    "langchain>=0.2.11",
    "langchain-openai>=0.1.21",
    "openai>=1.40.0",
]
